+++ addOns/soap/src/main/java/org/zaproxy/zap/extension/soap/ImportDialog.java
@@ -42,6 +42,7 @@ import org.zaproxy.addon.commonlib.ZapUriException;
 import org.zaproxy.zap.utils.FontUtils;
 import org.zaproxy.zap.utils.ThreadUtils;
 import org.zaproxy.zap.utils.ZapHtmlLabel;
+import org.zaproxy.zap.utils.ZapNumberSpinner;
 import org.zaproxy.zap.view.LayoutHelper;
 
 @SuppressWarnings("serial")
@@ -51,6 +52,7 @@ public class ImportDialog extends AbstractDialog {
 
     private final ExtensionImportWSDL extSoap;
     private JTextField fieldWsdl;
+    private ZapNumberSpinner fieldMaxMessages;
     private JButton buttonChooseFile;
     private JButton buttonCancel;
     private JButton buttonImport;
@@ -78,6 +80,13 @@ public class ImportDialog extends AbstractDialog {
         fieldsPanel.add(
                 getChooseFileButton(),
                 LayoutHelper.getGBC(2, fieldsRow, 1, 0.5, new Insets(0, 4, 4, 0)));
+        fieldsRow++;
+        fieldsPanel.add(
+                new JLabel(Constant.messages.getString("soap.importDialog.labelMaxMessages")),
+                LayoutHelper.getGBC(0, fieldsRow, 1, 0.5, new Insets(4, 0, 4, 4)));
+        fieldsPanel.add(
+                getMaxMessagesField(),
+                LayoutHelper.getGBC(1, fieldsRow, 2, 0.5, new Insets(4, 4, 4, 0)));
 
         int row = 0;
         add(fieldsPanel, LayoutHelper.getGBC(0, row, 2, 1.0, new Insets(8, 8, 4, 8)));
@@ -112,9 +121,10 @@ public class ImportDialog extends AbstractDialog {
             return false;
         }
 
+        int maxMessages = getMaxMessagesField().getValue();
         try {
             UriUtils.isValid(wsdlLocation);
-            extSoap.extUrlWSDLImport(wsdlLocation);
+            extSoap.extUrlWSDLImport(wsdlLocation, maxMessages);
             return true;
         } catch (ZapUriException e) {
             // Not a valid URI, try to import as a file
@@ -129,7 +139,7 @@ public class ImportDialog extends AbstractDialog {
                     });
             return false;
         }
-        extSoap.fileUrlWSDLImport(file);
+        extSoap.fileUrlWSDLImport(file, maxMessages);
         return true;
     }
 
@@ -149,6 +159,13 @@ public class ImportDialog extends AbstractDialog {
             setContextMenu(fieldWsdl);
         }
         return fieldWsdl;
+    }
+
+    private ZapNumberSpinner getMaxMessagesField() {
+        if (fieldMaxMessages == null) {
+            fieldMaxMessages = new ZapNumberSpinner(0, 0, Integer.MAX_VALUE);
+        }
+        return fieldMaxMessages;
     }
 
     private JButton getChooseFileButton() {
@@ -248,10 +265,12 @@ public class ImportDialog extends AbstractDialog {
 
         getImportButton().setEnabled(!show);
         getWsdlField().setEnabled(!show);
+        getMaxMessagesField().setEnabled(!show);
         getChooseFileButton().setEnabled(!show);
     }
 
     void clearFields() {
         getWsdlField().setText("");
+        getMaxMessagesField().changeToDefaultValue();
     }
 }