+++ addOns/fuzz/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+
+
+## [13.17.0] - 2026-08-07
 ### Changed
 - Update minimum ZAP version to 2.17.0.
 - Update dependency.
@@ -235,6 +238,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First version.
 
+[13.17.0]: https://github.com/zaproxy/zap-extensions/releases/fuzz-v13.17.0
 [13.16.0]: https://github.com/zaproxy/zap-extensions/releases/fuzz-v13.16.0
 [13.15.0]: https://github.com/zaproxy/zap-extensions/releases/fuzz-v13.15.0
 [13.14.0]: https://github.com/zaproxy/zap-extensions/releases/fuzz-v13.14.0