+++ addOns/llm/CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+
+## [0.0.1] - 2026-08-07
+
 ### Added
 - Basic stats
 - Support for Ollama, Azure OpenAI, Google Gemini, Claude (Anthropic), and OpenAI Compatible endpoints.
 - Integration points for other add-ons.
 - A tabbed LLM Chat panel.
 - Integration with the MCP add-on.
+
+[0.0.1]: https://github.com/zaproxy/zap-extensions/releases/llm-v0.0.1