+++ addOns/selenium/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 
+## [15.54.0] - 2026-08-12
+### Changed
+- Update Selenium to version 4.47.0.
+
 ## [15.53.0] - 2026-07-14
 ### Changed
 - Update Selenium to version 4.46.0.
@@ -384,6 +388,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated to Selenium 2.45 and moved to release
 
+[15.54.0]: https://github.com/zaproxy/zap-extensions/releases/selenium-v15.54.0
 [15.53.0]: https://github.com/zaproxy/zap-extensions/releases/selenium-v15.53.0
 [15.52.0]: https://github.com/zaproxy/zap-extensions/releases/selenium-v15.52.0
 [15.51.0]: https://github.com/zaproxy/zap-extensions/releases/selenium-v15.51.0