+++ addOns/soap/CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+
+
+## [32] - 2026-08-12
 ### Added
-- Allow Automation Framework users to limit the number of SOAP messages to import (`maxMessages`). Ex: If testing authentication, access, etc.
+- Allow users to limit the number of SOAP messages to import (`maxMessages`). Ex: If testing authentication, access, etc.
 
 ### Changed
 - Maintenance changes.
@@ -218,6 +221,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First version
 
+[32]: https://github.com/zaproxy/zap-extensions/releases/soap-v32
 [31]: https://github.com/zaproxy/zap-extensions/releases/soap-v31
 [30]: https://github.com/zaproxy/zap-extensions/releases/soap-v30
 [29]: https://github.com/zaproxy/zap-extensions/releases/soap-v29