+++ addOns/webdrivers/webdriverlinux/CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 
+## [216] - 2026-08-12
+### Changed
+- Update ChromeDriver to 151.0.7922.137.
+
+## [215] - 2026-08-07
+### Changed
+- Update ChromeDriver to 151.0.7922.108.
+
+## [214] - 2026-08-04
+### Changed
+- Update ChromeDriver to 151.0.7922.75.
+
 ## [213] - 2026-07-29
 ### Changed
 - Update ChromeDriver to 151.0.7922.71.
@@ -885,6 +897,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First release: Firefox v0.13.0 Chrome v2.27
 
+[216]: https://github.com/zaproxy/zap-extensions/releases/webdriverlinux-v216
+[215]: https://github.com/zaproxy/zap-extensions/releases/webdriverlinux-v215
+[214]: https://github.com/zaproxy/zap-extensions/releases/webdriverlinux-v214
 [213]: https://github.com/zaproxy/zap-extensions/releases/webdriverlinux-v213
 [212]: https://github.com/zaproxy/zap-extensions/releases/webdriverlinux-v212
 [211]: https://github.com/zaproxy/zap-extensions/releases/webdriverlinux-v211