+++ addOns/webdrivers/webdriverwindows/CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 
+## [217] - 2026-08-12
+### Changed
+- Update ChromeDriver to 151.0.7922.137.
+
+## [216] - 2026-08-07
+### Changed
+- Update ChromeDriver to 151.0.7922.108.
+
+## [215] - 2026-08-04
+### Changed
+- Update ChromeDriver to 151.0.7922.75.
+
 ## [214] - 2026-07-29
 ### Changed
 - Update ChromeDriver to 151.0.7922.71.
@@ -889,6 +901,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First release: Firefox v0.13.0 Chrome v2.27 IE 3.0.0
 
+[217]: https://github.com/zaproxy/zap-extensions/releases/webdriverwindows-v217
+[216]: https://github.com/zaproxy/zap-extensions/releases/webdriverwindows-v216
+[215]: https://github.com/zaproxy/zap-extensions/releases/webdriverwindows-v215
 [214]: https://github.com/zaproxy/zap-extensions/releases/webdriverwindows-v214
 [213]: https://github.com/zaproxy/zap-extensions/releases/webdriverwindows-v213
 [212]: https://github.com/zaproxy/zap-extensions/releases/webdriverwindows-v212